Refactor: move dirtab.h|cc to path_filters/
