+++ cvmfs/swissknife_pull.cc
@@ -23,13 +23,13 @@
 
 #include "atomic.h"
 #include "catalog.h"
-#include "dirtab.h"
 #include "download.h"
 #include "hash.h"
 #include "history_sqlite.h"
 #include "logging.h"
 #include "manifest.h"
 #include "manifest_fetch.h"
+#include "path_filters/dirtab.h"
 #include "signature.h"
 #include "smalloc.h"
 #include "upload.h"
